requirements.txt
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
kiwisolver==1.4.5
matplotlib==3.8.4
mysql-connector-python==8.4.0
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
PyQt6==6.7.0
PyQt6-Qt6==6.7.0
PyQt6-sip==13.6.0
python-dateutil==2.9.0.post0
six==1.16.0
bcrypt==4.2.1
pandas
prophet